// moonbowcl/src/app/model/dto/area-ruolo.ts
/*
* SPDX-FileCopyrightText: (C) Copyright 2023 C.S.I. Piemonte
*
* SPDX-License-Identifier: EUPL-1.2 */

export class AreaRuolo {
    idArea: number;
    codiceArea: string;
    nomeArea: string;
    idRuolo: number;
    codiceRuolo: string;
    nomeRuolo: string;
    dataUpdAbilitazione: Date;
    attoreUpdAbilitazione: string;
}